NEW SCREENS: stats with a chart,  add an expense with category creation popup
Stats screen uses the plugin Fl_Chart to integrate a bar diagram
- stat_screen.dart, dart.dart

The plus + button on the home screen now leads to add_expense.dart screen where the user can input an expense
A popup modal is also created if the user wants to create a new category.
It contains fields for category name, icon and icon bg colour.

Added 3 packages and the assets/icons/ directory in pubspec.yaml